requirements.txt
aiofiles
aiohttp
asyncio
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
motor
numpy
pillow
psutil
py-tgcalls==0.9.7
pykeyboard
pyrogram==2.0.106
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
unidecode
yt-dlp
youtube-search
youtube-search-python